requirements.txt
wheel
black
flake8
python-dotenv==0.21.0
selenium
requests
